requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
openai
faiss-cpu==1.7.4
PyPDF2==3.0.1
pdfplumber==0.10.2
python-dotenv==1.0.1
sentence-transformers==2.7.0
requests==2.31.0
gunicorn==22.0.0
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.16.3
python-dotenv
sentencepiece
transformers
accelerate

